requirements-nuinsseg.txt
numpy==1.26.4
scipy==1.13.0
scikit-image==0.22.0
scikit-learn==1.5.1
matplotlib==3.9.0
PyYAML==6.0.1
tqdm==4.66.4
Pillow==10.3.0
torch>=2.3.0
torchvision>=0.18.0
monai==1.3.1
tensorboardX==2.6.2.2
segment-anything==1.0
safetensors==0.4.3
timm==1.0.3
icecream==2.1.3
einops==0.8.0
opencv-python-headless==4.9.0.80
